Fix: atof para o parâmetro epsilon

+++ dp_example/blackjack.cpp
@@ -374,7 +374,7 @@ int main(int argc, char* argv[])
   const Natural num_episodes = atoi(argv[3]);
   const Natural min_batches = atoi(argv[4]);
   const Natural num_points = atoi(argv[5]);
-  const Natural epsilon = atoi(argv[6]);
+  const Natural epsilon = atof(argv[6]);
   const Natural inc_batches = (double) (num_batches - min_batches) / (double) num_points + 1;
 
   srand(run);
